master

+++ app/Http/Controllers/ProdutoController.php
@@ -20,74 +20,37 @@ class ProdutoController extends Controller
 
         $this->middleware('auth');
     }
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
-        $produtos = Produto::where('status',1)->orderBy('tipo')->paginate(15);
+        $produtos = Produto::where('status', 1)->orderBy('tipo')->paginate(5);
         return view('produto.list', ['produtos' => $produtos]);
     }
-     /**
-     * Display a listing of the inactive resource .
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function inactive()
     {
-        $produtos = Produto::where('status',0)->orderBy('nome')->paginate(15);
+        $produtos = Produto::where('status', 0)->orderBy('nome')->paginate(15);
         return view('produto.list', ['produtos' => $produtos]);
-
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         return view('produto.add');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request\StoreUpdateExtraRequest  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(StoreUpdateProdutoRequest $request)
     {
 
         $data = $request->all();
-        if ($request->hasFile('imagem') && $request->imagem->isValid()){
+        if ($request->hasFile('imagem') && $request->imagem->isValid()) {
             $path = $request->imagem->store('produtos');
             $data['imagem'] = $path;
         }
         Produto::create($data);
         Session::flash('mensagem_sucesso', 'Produto cadastrado com sucesso!');
-        return redirect()->route('produto.index'); 
+        return redirect()->route('produto.index');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
         $produto = Produto::where('id', $id)->first();
@@ -96,43 +59,38 @@ class ProdutoController extends Controller
         }
         return view('produto.add', ['produto' => $produto]);
     }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+    
     public function update(StoreUpdateProdutoRequest $request, $id)
     {
-        if (!$produto = $this->repository->find($id)){
-            return redirect()->back();
-        }
-        $data = $request->all();      
-        
-        if ($request->hasFile('imagem') && $request->imagem->isValid()){
-            if ($produto->imagem && Storage::exists($produto->imagem)) {
-                Storage::delete($produto->imagem);
+        $status = Produto::where('id', $id)->first();
+        if($status->status === 1){
+            if (!$produto = $this->repository->find($id)) {
+                return redirect()->back();
             }
-            $path = $request->imagem->store('produto');
-            $data['imagem'] = $path;
-        }
-        $produto->update($data);
-        Session::flash('mensagem_sucesso', 'Produto editado com sucesso!');
-        return redirect()->route('produto.index');   
-    }
+            $data = $request->all();
+    
+            if ($request->hasFile('imagem') && $request->imagem->isValid()) {
+                if ($produto->imagem && Storage::exists($produto->imagem)) {
+                    Storage::delete($produto->imagem);
+                }
+                $path = $request->imagem->store('produto');
+                $data['imagem'] = $path;
+            }
+            $produto->update($data);
+            Session::flash('mensagem_sucesso', 'Produto editado com sucesso!');
+            return redirect()->route('produto.index');
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+        }        
+        else {
+            Session::flash('mensagem_alerta', 'Somente produtos ativos podem ser editados!');
+            return redirect()->route('produto.index');
+        }
+    }
+    
     public function destroy($id)
     {
         $produto = $this->repository->where('id', $id)->first();
-        if(!$produto){
+        if (!$produto) {
             return redirect()->back();
         }
         if ($produto->imagem && Storage::exists($produto->imagem)) {
